+++ examples/simpleble/android/app/src/main/java/org/simpleble/android/Peripheral.kt
@@ -1,0 +1,13 @@
+package org.simpleble.android
+
+class Peripheral {
+
+    private var instanceId: Long = -1
+    private var adapterId: Long = -1
+
+    constructor(newAdapterId: Long, newInstanceId: Long) {
+        this.adapterId = newAdapterId
+        this.instanceId = newInstanceId
+    }
+
+}